Make `.main-content` Play Nicely with Alli Theme
Removing the `container-fixed` from main content makes it responsive.
Give it a little bit of padding on the top and bottom so we aren't
butting right up against the menu.

The container fluid padding change makes the content line up with the
top menu items. This plays nicely with the row gutters and such.

Also remove a mixin that was not necessary.